requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=2.2.0
argon2-cffi=20.1.0
async_generator=1.10
attrs=20.3.0
babel=2.9.0
backcall=0.2.0
backports=1.0
backports.functools_lru_cache=1.6.4
blas=2.106
bleach=3.3.0
brotlipy=0.7.0
ca-certificates=2020.12.5
certifi=2020.12.5
cffi=1.14.5
chardet=4.0.0
colorama=0.4.4
cryptography=3.4.7
cudatoolkit=11.1.1
cycler=0.10.0
decorator=5.0.7
defusedxml=0.7.1
deprecation=2.1.0
entrypoints=0.3
freetype=2.10.4
icu=58.2
idna=2.10
importlib-metadata=4.0.1
intel-openmp=2021.2.0
ipykernel=5.5.3
ipython=7.22.0
ipython_genutils=0.2.0
jedi=0.18.0
jinja2=2.11.3
jpeg=9b
json5=0.9.5
jsonschema=3.2.0
jupyter-packaging=0.9.2
jupyter_client=6.1.12
jupyter_core=4.7.1
jupyter_server=1.6.4
jupyterlab=3.0.14
jupyterlab_pygments=0.1.2
jupyterlab_server=2.4.0
kiwisolver=1.3.1
libblas=3.9.0
libcblas=3.9.0
libflac=1.3.3
liblapack=3.9.0
liblapacke=3.9.0
libogg=1.3.4
libopus=1.3.1
libpng=1.6.37
libsndfile=1.0.31
libsodium=1.0.18
libtiff=4.1.0
libuv=1.41.0
libvorbis=1.3.7
llvmlite=0.36.0
lz4-c=1.9.3
m2w64-gcc-libgfortran
m2w64-gcc-libs
m2w64-gcc-libs-core
m2w64-gmp
m2w64-libwinpthread-git
markupsafe=1.1.1
matplotlib=3.4.1
matplotlib-base=3.4.1
mistune=0.8.4
mkl=2020.4
msys2-conda-epoch
nbclassic=0.2.7
nbclient=0.5.3
nbconvert=6.0.7
nbformat=5.1.3
nest-asyncio=1.5.1
ninja=1.10.2
notebook=6.3.0
numba=0.53.1
numpy=1.20.2
olefile=0.46
openssl=1.1.1k
packaging=20.9
pandoc=2.13
pandocfilters=1.4.2
parso=0.8.2
pesq
pickleshare=0.7.5
pillow=8.2.0
pip=21.1
prometheus_client=0.10.1
prompt-toolkit=3.0.18
pycparser=2.20
pydub=0.25.1
pygments=2.8.1
pyopenssl=20.0.1
pyparsing=2.4.7
pypesq
pyqt=5.9.2
pyrsistent=0.17.3
pysocks=1.7.1
pysoundfile=0.10.3.post1
pystoi=0.3.3
python=3.8.8
python-dateutil=2.8.1
python_abi=3.8
pytorch=1.8.1
pytz=2021.1
pywin32
pywinpty
pyzmq=22.0.3
qt=5.9.7
requests=2.25.1
scipy=1.6.2
send2trash=1.5.0
setuptools=49.6.0
sip=4.19.13
six=1.15.0
sniffio=1.2.0
sqlite=3.35.5
terminado=0.9.4
testpath=0.4.4
tk=8.6.10
tomlkit=0.7.0
torchaudio=0.8.1
torchvision=0.9.1
tornado=6.1
tqdm=4.60.0
traitlets=5.0.5
typing_extensions=3.7.4.3
urllib3=1.26.4
vc
vs2015_runtime
wcwidth=0.2.5
webencodings=0.5.1
wheel=0.36.2
win_inet_pton=1.1.0
wincertstore=0.2
winpty
xz=5.2.5
zeromq=4.3.4
zipp=3.4.1
zlib=1.2.11
zstd=1.4.9
